Mark data as stale on refresh instead of clearing cache
